fix: Stuff 와 Post 도메인 병합으로 인한 Dibs 메서드 수정

+++ src/main/java/com/sparta/wuzuzu/domain/dibs/service/DibsService.java
@@ -47,6 +47,11 @@ public class DibsService {
         Long postId
     ) {
         Dibs dibs = dibsRepository.findByUserAndPostId(user, postId);
+
+        if(dibs == null){
+            throw new IllegalArgumentException("Dibs is empty.");
+        }
+
         dibsRepository.delete(dibs);
     }
 }
